+++ resources/meta/3rd/UniComm/library/io.d.lua
@@ -1,8 +1,11 @@
 ---@meta
 
 ---
+---Extends the standard Lua I/O library with UniComm application operations.
 ---Relative paths used by `io.open`, `io.input`, `io.output`, `io.lines`,
 ---`os.remove`, and `os.rename` start from the current workspace.
+---
+---[IO demo](../demo/io.lua)
 ---
 io = {}